Module-11-LangChain-Framework/05_Embedding_Vectorstore/requirements.txt
gradio
python-dotenv
langchain-community
langchain-openai
langchain-huggingface
langchain-chroma
numpy
openai
chromadb
